+++ Forge.Forms/src/Forge.Forms.Demo/Models/TextElements.cs
@@ -1,0 +1,12 @@
+﻿using Forge.Forms.Annotations;
+
+namespace Forge.Forms.Demo.Models
+{
+    [Title("Title")]
+    [Heading("Heading")]
+    [Text("Text")]
+    [ErrorText("ErrorText")]
+    public class TextElements
+    {
+    }
+}