+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 # Changelog
 
+## 1.0.16
+
+- Added error text element.
+    - You can add it via `[ErrorText]` attribute in classes.
+    - You can add it via `<error>` tag in XML.
+
 ## 1.0.15
 
-- Added parameterized value converters https://github.com/WPF-Forge/Forge.Forms/commit/fd1ecb0873b10d81a6ce479557317e5167a43664
-- Non-input XML elements accept `visible` property https://github.com/WPF-Forge/Forge.Forms/commit/42e23e238e9c909baa8d6a48f10cff9897dc9daa
+- Added parameterized value converters which are created from factories registered in `Resource` class.
+- Non-input XML elements accept `visible` property.
+- Added `Invalidate`, `IsValid`, `ValidateWithoutUpdate`, `GetValidationErrors` utilities in ModelState class.
+    - `Invalidate` temporarily fails validation for a property. If you modify the field, call `Validate`, or call `ValidateWithoutUpdate`, the marked error will disappear. To check the state of this validation error use `IsValid`/`GetValidationErrors`.
+    - `IsValid` is useful if you want to check current validation state without triggering side-effects like `Validate`, which sometimes produces unexpected results.
+    - `ValidateWithoutUpdate` should generally be avoided unless you are dealing with some strange edge cases.
+    - `GetValidationErrors` returns the list of error strings from specified properties/whole model.
+- Added `Must.BeInvalid` (and `Must.Fail` alias) validator for external validation. This validator always fails, so you must combine it with the `When` condition.
+- Added `ValueAttribute.OnActivation` and `ValueAttribute.OnDeactivation` validation actions which specify what happens when the validator is enabled/disabled from the `When` condition.
+- Added `SelectionType.RadioButtonsInline` for horizontal layout of radio buttons in selections.
 
 ## 1.0.14
 
-- Stable NuGet release
+- Stable NuGet release.